CWndStateConfirm: remove all the packets related to increasing by one the stats
It was probably legacy code as CWndStateConfirm::SendYes is only
used in one place: when pressing Enter or Escape with the window
open. In this case, it was sent to increase the stat with the id 0,
and WorldServer was early returning without doing anything.

In this commit, we remove all the functions that were related to
increasing by one statistics as they were never used.