[#57] Player's position does not go through LoginServer anymore
When loading the list of characters of an account, database server
was also loading and sending the current world and the position of the
characters.

Neuz received this information from LoginServer, and sent to CacheServer
the world (it thinks it will spawn in).

I think it is a waste of energy because:
- The world of a player is never updated by CacheServer, and it is barely used.
- It is only used to get the right WorldServer to spawn the player in. But as
all worldservers are supposed to have all valid worlds it is kind of pointless
- The fact that a world id is in the list of juridictions of a server does not mean
that the server is question has successfully loaded the world. Moreover, the world
id the client send and the real world may be mismatched so the security benefit
is very limited.

This change can go against two kind of possible modifications:
- Displaying in the character list the world a player is in.
- Some modifications where Cache is always aware of where a player is, and can change
the WorldServer a player is in depending of its current world.

The first point is very easy to undo, and will be easier to implement when
PACKETTYPE_PLAYER_LIST will use a proper struct
The second point requires to have a deep understanding of the architecture to implement,
so users that want to implement this will not be bothered by this change.