[#57] Remove PacketType Passage
- Why is it useless?
It allowed a worldserver to pass a message to a precise WorldServer

It is never used in V15, never used in V21.
I can hardly see a context where we want to use it.

The only cases I thought of are:
- A wants to send a message to the world where the player P is. Core server
should find by itself where the player is.
- A wants to relocate a player. It should senda a message to CacheServer
that will be in charge of the switch instead.

- Why do I remove this feature?
It uses the GetWorldSrvrDPID function that I want to rework to only
take the server id of multi to return the DPID.