Emmanuel Macron approuve cet engagement
You could open the bank by sacrificing any item as the game would not check that the item you used is actually a scroll of custody / scroll to open now the bank. It is now fixed. And the joke is that Manu is a former M&A banker, and engagement is the worst possible translation for commit