CEventMonster
- Remove the SetEventMonster that sets an iterator so you can read
the information you want idea. It is bad design wise because we
prefer const
- As official seems to have < 10 event monsters, use a flat map
- I had an idea for the global instance without singleton thing:
use a static local variable of an actual instance, and define
static methods as a proxy for the actual instance. I like the fact
that to call the methods, we just have to use the name of the class
(like a namespace) and the fact that we still have a proper class
for the implementation.