[Spawn] Bosses properly respawn even if players are trolling
There is a latent bug in FlyFF where you have to watch something
disappear to let it respawn.

Usually it does not matter, but there are two scenarii where it
is annoying:
- In servers that decide to eject 50km away monsters when they are
killed.
- For giant hunting, people may kill the giant and teleport away
or run away very fast

For the first case, don't implement this feature tyvm.

The second case is the one addressed by this commit: in this
commit, when a monster dies, we check if they are the only one
in the spawner. If it is the case, we signal to the spawner that
the monster died now, instead of waiting the proper removal of
the monster.

So now, if a player does not wait for the boss death, in the worst
case, the player that comes next will see both the previous giant
dying and the new one sitting here close to its predecessor.

Side effect: bosses may respawn few seconds earlier than before. But
as the timer is expected to be long (at least one minute), we do
not care. That's also why we only do this for spawner with only one
thing to spawn : we do not want to speed up everything