Post: Get rid of second parmeter in CMail(Box)?::Serialize
The only case were it was set to FALSE was trough CPost::Serialize,
which was only called with FALSE as its second parameter.

Fully Implement CPost::operator<< / operator>> considering that
the bData parameter is only false, and consider that the second
parameter is always true.

As CPost::operator<</>> does not send/receive everything,
we use some kind of proxy to warn the user about the fact not
everything will be sent.