Merci Chujin
It some cases, multiple guilds with the same name can participate
in the same guild combat: they simply need to have an empty name.

Changing the way the guild ranking is stored to use guild ids
instead of names.

(Chujin is a former FlyFF player who used to participate in GC
alone in an unnamed guild)