Fix mem leak related to animated wings, found by Florist
== The memleak
Post by Florist on Lodelight describing the leak and a possible fix:
https://lodelight.com/threads/memory-leak-abuse-method-fix.671/

== The fix used there
Here, we use a different approach by introducing the concept of
"unenforced maybe owned ptr" (unenforced MOP) for CModels and then
casting the unenforced MOP into an actual MOP.

The unenforced MOP being the ModelPointerWithOwnershipInfo class.

The unenforced MOP version is used to avoid having to adapt the
code everywhere else in the code.

In future commits, the ModelPointerWithOwnershipInfo class will
evolve further