coan source -U__CERTIFIER_COLLECTING_SYSTEM --filter c,h,cpp --recurse --no-transients .
This define is related to a system to collect to IP address a neuz
connects to.

For a private server, there is absolutely no point in using such
system. For the official server, it may have been useful when only
server executables leaked, but now, when there is a leak, the
whole source leaks so it is even pointless for them. The proof
being that the v21 code removed this sytem.