Remove multiple cache support in LoginSrvr
- It was not really supported
LoginServer is supposed to connect to one Core.
Core can only receive connections from one Cache,
see Source/CoreServer/DPCacheSrvr.cpp l102-127

1 Login = 1 Core. 1 Core = 1 Cache. By transitivity,
1 Login = 1 Cache.

Using the multiple cache per login feature would lead to a buggy
behaviour. Hence, removing this feature is not a regression
but a bug proofing.

- There are no reason to have multiple login per Cache

The implemented cache mechanic was simply dispatching players on
caches. Cache is a very low weight component, and sharing it
between all players in the same region has no sense.

The only reasonnable use case is if you want to have cache in
multiple regions of the world. For example a Cache in Europe
and a cache in NA to lower ping. In such use case, you probably
want to connect the cache to worlds in the same region of the world.
The implemented Cache mechanic was useless for this purpose, as
in this model, you send the cache IP address of the selected world.

That should be Certifier's job anyway.

I still do not understand why Login and Cache Server as separated,
but this is another discussion (#58).