Remove functions that may change the device
Changing the device is not supported at all in the source code:
the g_Neuz.m_pd3dDevice variable leaks everywhere without any
real strategy to be able to change all instances and the allocated
resources.

While it would be nice to debug it, let's be honnest, it would
be very long for limited benefits.

This very opinionated commit, and the following ones, are going
to remove the illusion that g_Neuz.m_pd3dDevice is not
considered fixed for the whole duration of the process.

The expected benefits are:
- Simplification of some functions, in particular the AddTexture
function that has a complex API because of g_Neuz.m_pd3dDevice.
- By removing the impossible to debug, it will be possilble to
actual focus on what is more easily fixable and is more important,
i.e the Initialize/Restore/Delete/Invalidate/DeviceObject functions.