Remove all D3DDevice parameters and attributes
**This is a very opinionated choice.**

TL;DR: Simpler APIs, stop passing the device in every other function

== Let us start with two facts:
- Changing the device is not really supported. You can observe it by
calling the code related to IDM_CHANGEDEVICE. Changing the device
crashes the game because no consideration for such feature has been
made.
- Debugging the device change would require a lot of modifications
for very little benefit. To the best of my knowledge, noone has done
that and nobody asks for this feature.

== The device is stored everywhere for nothing
In FlyFF code, the device leaks everywhere, is locally stored everywhere
etc. But unless you call the code related to IDM_CHANGEDEVICE, and
some other never actually used functions (like the one to change
the current process to fullscreen mode), the D3DDevice should never
change.

Moreover, the choice between D3DDEVICE, a pd3dDevice parameter, an
m_pd3dDevice local field, or even other possibilities (p2DRender local
attribute, through CWndBase::m_pApp) is sometimes inconsistent in the
same function. For example, see the `CLandscape::RenderWater` function
that uses both `pd3dDevice` and `m_pd3dDevice`.

== Some APIs are made more complicated for no reason
Some functions that are very used very often, like the AddTexture
function, are annoying to use because you need to pass the device. If
you did not study the (lack of) differences between the D3DDEVICEs, you
are never really sure which one to pass. Turns out it does not matter
because they are all the same.

== This commit
In this commit, all LPDIRECT3DDEVICE9 parameter and fields are removed.
The only ones that are kept are:
- the field in CD3DApplication, as it is where it is originally stored
- In some D3DX utility functions, are they are inbetween FlyFF code
and the Windows API.

Some (not so) global variables named m_pd3dDevice and pd3dDevice are
providing compatibility with existing code, in particular the Render
functions that heavily use these variables. While unifying
D3DDEVICE, pd3dDevice and m_pd3dDevice would be nice, this is left
for another commit because this one is already huge.