Hard code m_pWndRoot as g_WndMng
m_pWndRoot was a static field of CWndBase to tell which window is
the root.

Except in practice, the root is always the only CWndMgr instance
named g_WndMng, and it is used almost everywhere.

There are more occurrences of g_WndMng than m_pWndRoot, and the
root is considered fixed.

Definitely setting the root as g_WndMng clarifies the code.

Anyway, if the root should not be considered fixed, m_pWndRoot
should have been a simple field and not a static field.