Remove CWndNeuz::Initialize
Initialize is not a requirement for CWndNeuz, it is only a requirement
for some CWndMessageBox and applets. For all others CWndNeuz, it is
more of a convention than a requirement.

This commit removes the CWndNeuz::Initialize definition that was
dubious.

For applets, the only place where Initialize is used is when creating
the applet. In this commit, the Initialize part is moved to the
CWndNeuz derived class instantiation.

For CWndMessageBox, proper polymorphism is applied for derived classes,
with the introduction of the CWndCustomMessageBox classe for
derived classes that are then passed to the OpenMessageBox method.

For all other classes, the Initialize method is kept, but is not
overriding anything anymore. The pWndParent parameter is also kept
as part as the convention.

However, it may evolve further as many CWndNeuz classes are expected
to be the only one at any given time. This is left for another commit...