Remove GUILD_MSG_HEADER / PACKETTYPE_GUILD_MSG_CONTROL
This packet was only used on creating a Guild Cloak ... on paper.

In practice the packet was doing:
WorldServer -> CoreServer -> CacheServer -> Neuz
Instead of the probably intended World -> CoreServer -> All Worlds

In the indented (all) World(s) Handler, the amount of penya in
the guild is sent to all players. This task is already performed
in the functions that used to send a PACKETTYPE_GUILD_MSG_CONTROL.

-> This packet had no effect in practice and debugging it would
provide no extra service (the all worlds handler is actually
glitched as it sends to all members that the one that payed for
the cloak is themselves)

The GUILD_MSG_HEADER struct is hard to read, supposed to be
generic but only used once, so removing it clarifies the code