Replace all ForLinkMap occurrences with GetLandRange
"In the end, you just replaced two macros with two macros
and unreadable templates"

Some benefits of the GetLandRange macro over FOR_LAND:
- You can wite a ranged for loop, and you are not forced to name
the loop variable pObj. Also, the loop variable is initialized
when it is declared which is better.

- GetLandRange can serve a specialization of the CObj class,
no more casting from CObj to CUser when looping on
a LinkType::Player.

- In C++29, not allocating for generators that are directly used
will be standardized, so all the complicated iterators will be
replaced by a generator (but GetLandRange will probably remain
a macro sadly)

Benefits of the GetLandRange macro over ForLinkMap function:
- No lambda, no weird return parameter, an intuitive way to
escape from the loop
- Argument order is the same as FOR_LAND