Add .gitattributes for eol of some files
Usually, you should clone useless on a Windows OS with automatic
replacement of lf to crlf.

However, if you download the repo as a .zip from Github, or if you
work on Linux, or for whatever other reason, the changes may not
be performed.

This causes issue for files processed by CScanner as it expects
new lines to be crlf. This commit adds a gitattributes to force
all these systems to have the correct line ending for some files.

Thanks Tratox for reporting that there was an error and taking the
time to do my tests for investigation.