Add operator<</>>(CAr, std::map/vector)
These functions are convenient

But we still need to enforce the fact that we do not trust
players.

Solution:
- operator>> is only implemented in Neuz
- For other executables, there are ReadMap and ReadVector. These
functions expect a maximum length. If the user sends a map or a
vector with a higher size, std::nullopt is returned instead

Vector / Map concept is inspired by this post from Kia on
Lodelight:
https://lodelight.com/threads/sendawakening.857/#post-4313