Add CMiniGamePairgame in MiniGames.h
The prupose of this serie of commit is to put all rainbow
minigames in a single header-cpp files pair.

I consider that having tons of small classes in tons of small
separated files makes the comprehension harder: if the classes
are smalls and fit a similar purpose, they can be in the same
files.

The serie of commit does not address the fact that the minigames
are client side managed